@@ -46,7 +46,7 @@ class ElasticSearchRepository implements Repository
     /**
      * {@inheritDoc}
      */
-    public function save(Identifiable $data)
+    public function save(Identifiable $data): void
     {
         Assertion::isInstanceOf($data, $this->class);
 
@@ -66,7 +66,7 @@ class ElasticSearchRepository implements Repository
     /**
      * {@inheritDoc}
      */
-    public function find($id)
+    public function find($id): ?Identifiable
     {
         $params = [
             'index' => $this->index,
@@ -106,7 +106,7 @@ class ElasticSearchRepository implements Repository
     /**
      * {@inheritDoc}
      */
-    public function remove($id)
+    public function remove($id): void
     {
         try {
             $this->client->delete([
